ai-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.26.0
opencv-python>=4.8.0
pillow>=10.3.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
redis>=5.0.1
opencv-contrib-python>=4.8.0


